+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - None
 ### Changed
 - updated space export, import and clone functions in myems-api
-
+- updated microgrid reporting in myems-api and myems-web
+- 
 ### Fixed
 - fixed chart issues in myems-web